+++ R/deps.R
@@ -3,29 +3,34 @@
 #' @param pac character a package name.
 #' @param fields character vector with possible values `c("Depends", "Imports", "LinkingTo", "Suggests")`. Default: `c("Depends", "Imports", "LinkingTo")`
 #' @param lib.loc character vector. Is omitted for non NULL version., Default: NULL
-#' @param attr logical specify if a package and its version should be added as a attribute of data.frame or for FALSE as a additional record. Default: TRUE
 #' @param base logical if to add base packages too. Default: FALSE
 #' @param local logical if to use newest CRAN packages, where by default local ones are used. Default: TRUE
 #' @param description_v if the dependencies version should be taken from description files, minimal required. Default: FALSE
+#' @param attr logical specify if a package and its version should be added as a attribute of data.frame or for FALSE as a additional record. Default: TRUE
+#' @param recursive logical if to assess the dependencies recursively. Default: TRUE
+#' @param repos character the base URL of the repositories to use. Default `https://cran.rstudio.com/`
 #' @return data.frame with packages and their versions. Versions are taken from `installed.packages` or newest released.
 #' @export
 #' @examples
 #' pacs::pac_deps("stats", base = TRUE)$Package
 #' pacs::pac_deps("memoise")$Package
 #' pacs::pac_deps("memoise", description_v = FALSE)
-#'
+#' # raw dependencies from DESCRIPTION file
+#' pacs::pac_deps("memoise", description_v = TRUE, recursive = FALSE)
 pac_deps <- function(pac,
                      fields = c("Depends", "Imports", "LinkingTo"),
                      lib.loc = NULL,
                      base = FALSE,
                      local = TRUE,
                      description_v = FALSE,
-                     attr = TRUE) {
+                     attr = TRUE,
+                     recursive = TRUE,
+                     repos = "https://cran.rstudio.com/") {
   stopifnot((length(pac) == 1) && is.character(pac))
   stopifnot(all(fields %in% c("Depends", "Imports", "Suggests", "LinkingTo")))
   stopifnot(is.logical(base))
   stopifnot(is.logical(attr))
-  stopifnot(!all(c(!local, description_v)))
+  stopifnot(is.logical(recursive))
 
   if (local) {
     stopifnot(pac %in% c(rownames(utils::installed.packages(lib.loc = lib.loc)), pacs_base()))
@@ -51,7 +56,7 @@ pac_deps <- function(pac,
       for (r in res) {
         if (r != "R" && !r %in% paks_global) {
           paks_global <<- c(r, paks_global)
-          deps(r, fields)
+          if (recursive) deps(r, fields)
         }
       }
     }
@@ -59,28 +64,22 @@ pac_deps <- function(pac,
     deps(pac, fields)
     v_base <- utils::installed.packages(lib.loc = lib.loc)
   } else {
-    stopifnot(pac %in% rownames(available_packages()))
+    stopifnot(pac %in% rownames(available_packages(repos)))
     paks_global <- tools::package_dependencies(pac,
-      db = available_packages(),
+      db = available_packages(repos),
       which = fields,
-      recursive = TRUE
+      recursive = recursive
     )[[1]]
-    v_base <- available_packages()
+    v_base <- available_packages(repos)
     pac_v <- v_base[pac, c("Version")]
   }
 
   res <- unique(c(
-    setdiff(
-      paks_global,
-      c(
-        pac,
-        if (!base) {
-          c(pacs_base(), "R")
-        } else {
-          NULL
-        }
-      )
-    ),
+    if (base) {
+      c(paks_global, pacs_base())
+    } else {
+      setdiff(paks_global, c(pacs_base()))
+    },
     if (!attr) {
       pac
     } else {
@@ -89,7 +88,11 @@ pac_deps <- function(pac,
   ))
 
   if (description_v) {
-    res_df <- installed_descriptions(lib.loc, fields, c(res, pac))
+    if (local) {
+      res_df <- installed_descriptions(lib.loc, fields, if (recursive) c(res, pac) else pac)
+    } else {
+      res_df <- available_descriptions(repos, fields, if (recursive) c(res, pac) else pac)
+    }
     res_df <- res_df[res_df$Package %in% res, ]
   } else {
     res_df <- as.data.frame(v_base[res, c("Package", "Version"), drop = FALSE])