+++ R/size.R
@@ -50,31 +50,29 @@ pacs_size <- function(pacs = NULL, lib.loc = NULL) {
 #' @param pac character a package name.
 #' @param fields character vector, Default: `c("Depends", "Imports", "LinkingTo")`
 #' @param lib.loc character vector, Default: NULL
-#' @param base logical if to add base packages too. Default: FALSE
-#' @param exclude_joint integer exclude packages which are dependencies of at least N other packages. Default: 0
+#' @param exclude_joint integer exclude packages which are dependencies of at least N other packages, not count main package dependencies. Default: 0
+#' @note R base packages are not counted.
 #' @return numeric size in bytes, to get MB then divide by `10**6`.
 #' @export
 #' @examples
-#' # size in Mb
-#' pacs::pac_true_size("stats") / 10**6
+#' # size in Mb, with all its dependencies
+#' pacs::pac_true_size("memoise") / 10**6
 #' # exclude packages if at least one other package use it too
 #' pacs::pac_true_size("memoise", exclude_joint = 1L) / 10**6
-#'
 pac_true_size <- function(pac,
                           fields = c("Depends", "Imports", "LinkingTo"),
                           lib.loc = NULL,
-                          base = FALSE,
                           exclude_joint = 0L) {
   stopifnot(is.null(lib.loc) || all(lib.loc %in% .libPaths()))
   stopifnot(all(fields %in% c("Depends", "Imports", "LinkingTo", "Suggests")))
   stopifnot(is.integer(exclude_joint))
 
-  pacs_all <- pac_deps(pac, fields = fields, lib.loc = lib.loc, attr = FALSE, base = base)$Package
+  pacs_all <- pac_deps(pac, fields = fields, lib.loc = lib.loc, attr = FALSE, base = FALSE)$Package
 
   if (exclude_joint) {
-    depsy <- stats::setNames(lapply(pacs_all, function(x) tools::dependsOnPkgs(x)), pacs_all)
+    depsy <- stats::setNames(lapply(pacs_all, function(x) setdiff(tools::dependsOnPkgs(x), pacs_all)), pacs_all)
     pacs_all <- setdiff(pacs_all, names(Filter(function(x) length(x) > exclude_joint, depsy)))
   }
 
-  sum(pacs_size(setdiff(pacs_all, "R"), lib.loc = lib.loc))
+  sum(pacs_size(unique(c(pac, setdiff(pacs_all, "R"))), lib.loc = lib.loc))
 }