+++ R/valid.R
@@ -1,18 +1,21 @@
 #' Validate the library.
 #' @description Compare current and expected packages under `.libPaths`.
-#'Checking the healthy of the library, which packages are newest one.
+#' Checking the healthy of the library, which packages are newest one.
 #' Optionally added life duration of each package.
 #' @param lib.loc character. Default: NULL
 #' @param fields character vector with possible values `c("Depends", "Imports", "LinkingTo", "Suggests")`. Default: `c("Depends", "Imports", "LinkingTo")`
 #' @param lifeduration logical if to add life duration column, might take some time. Default: FALSE
-#' @return data.frame with 5/6 columns Package Version.expected.min Version.have. "" means newest version.
+#' @param checkred logical if to add R CRAN check page status, any WARNING or ERROR will give TRUE. Default FALSE
+#' @return data.frame with 5/7 columns Package Version.expected.min Version.have. "" means newest version.
 #' @note Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files.
+#' When turn on the `lifeduration` and/or `is_red` options, calculations might be time consuming.
 #' @export
 #' @examples
 #' lib_validate()
 lib_validate <- function(lib.loc = NULL,
                          fields = c("Depends", "Imports", "LinkingTo"),
-                         lifeduration = FALSE) {
+                         lifeduration = FALSE,
+                         checkred = FALSE) {
   stopifnot(is.null(lib.loc) || all(lib.loc %in% .libPaths()))
   stopifnot(all(fields %in% c("Depends", "Imports", "Suggests", "LinkingTo")))
   stopifnot(is.logical(lifeduration))
@@ -37,9 +40,17 @@ lib_validate <- function(lib.loc = NULL,
 
   result <- result[!is.na(result$Package) & !(result$Package %in% c("NA", pacs_base())), ]
 
-  if (lifeduration) result$life_duration <- apply(result, 1, function(x) pac_lifeduration(x["Package"], x["Version.have"]))
+  if (lifeduration) {
+    cat("Please wait, Packages life durations are assessed.\n")
+    result$life_duration <- apply(result, 1, function(x) pac_lifeduration(x["Package"], x["Version.have"]))
+  }
 
   result$newest <- apply(result, 1, function(x) is_last_release(x["Package"], x["Version.have"]))
+
+  if (checkred) {
+    cat("Please wait, Packages CRAN check statuses are assessed.\n")
+    result$checkred <- vapply(seq_len(nrow(result)), function(x) result$newest[x] && pac_checkred(result$Package[x]), logical(1))
+  }
 
   result
 }
@@ -49,7 +60,7 @@ lib_validate <- function(lib.loc = NULL,
 #' @param pac character a package name.
 #' @param lib.loc character. Default: NULL
 #' @param fields character vector with possible values `c("Depends", "Imports", "LinkingTo", "Suggests")`. Default: `c("Depends", "Imports", "LinkingTo")`
-#' @return data.frame with 5 columns Package Version.expected.min Version.have. "" means newest version.
+#' @return data.frame with 7 columns Package Version.expected.min Version.have. "" means newest version.
 #' @note Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files.
 #' @export
 #' @examples
@@ -77,6 +88,8 @@ pac_validate <- function(pac, lib.loc = NULL, fields = c("Depends", "Imports", "
 
   result$newest <- apply(result, 1, function(x) is_last_release(x["Package"], x["Version.have"]))
 
+  result$is_redcheck <- apply(result, 1, function(x) pac_checkred(x["Package"]))
+
   result
 }
 
@@ -85,12 +98,11 @@ pac_validate <- function(pac, lib.loc = NULL, fields = c("Depends", "Imports", "
 #' @param pacs character vector packages names.
 #' @param lib.loc character. Default: NULL
 #' @param fields character vector with possible values `c("Depends", "Imports", "LinkingTo", "Suggests")`. Default: `c("Depends", "Imports", "LinkingTo")`
-#' @return data.frame with 5 columns Package Version.expected.min Version.have. "" means newest version.
+#' @return data.frame with 7 columns Package Version.expected.min Version.have. "" means newest version.
 #' @note Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files.
 #' @export
 #' @examples
 #' pacs_validate(c("memoise", "rlang"))
-#'
 pacs_validate <- function(pacs, lib.loc = NULL, fields = c("Depends", "Imports", "LinkingTo")) {
   stopifnot(is.null(lib.loc) || all(lib.loc %in% .libPaths()))
   stopifnot(all(fields %in% c("Depends", "Imports", "Suggests", "LinkingTo")))
@@ -105,6 +117,16 @@ pacs_validate <- function(pacs, lib.loc = NULL, fields = c("Depends", "Imports",
     all = TRUE,
     suffix = c(".expected.min", ".have")
   )
+
+  result$version_status <- apply(result, 1, function(x) utils::compareVersion(x["Version.have"], x["Version.expected.min"]))
+
+  result <- result[!is.na(result$Package) & !(result$Package %in% c("NA", pacs_base())), ]
+
+  result$life_duration <- vapply(result$Package, pac_lifeduration, numeric(1))
+
+  result$newest <- apply(result, 1, function(x) is_last_release(x["Package"], x["Version.have"]))
+
+  result$is_redcheck <- apply(result, 1, function(x) pac_checkred(x["Package"]))
 
   result
 }